chore(cosmic-swingset): Rename maybe-lockdown.js to maybe-unsafe-lockdown.js
...for better indicating its behavior
