+++ packages/cosmic-swingset/src/helpers/maybe-unsafe-lockdown.js
